Clean up module structure
